refactor: reorganize into packlets
