+++ src/packlets/app/AppHeader.tsx
@@ -1,4 +1,5 @@
 import { ComponentChildren } from 'preact'
+
 import { logo } from '../../packlets/branding'
 
 export default function AppHeader(props: { rightContent?: ComponentChildren }) {
