+++ src/packlets/localize/t.tsx
@@ -1,4 +1,4 @@
-import { LocalizedText } from '../components/LocalizedText'
+import { LocalizedText } from './LocalizedText'
 
 export function t(th: string, en: string) {
   return <LocalizedText th={th} en={en} />
