+++ src/packlets/qrcode/QRCode.tsx
@@ -1,9 +1,7 @@
 import { FunctionalComponent } from 'preact'
 import { useEffect, useRef } from 'preact/hooks'
 import * as qr from 'qrcode'
-import createPixelsRenderer, {
-  PixelPosition,
-} from '../services/createPixelsRenderer'
+import createPixelsRenderer, { PixelPosition } from './createPixelsRenderer'
 
 type QRCodeProps = {
   payload: string | qr.QRCodeSegment[]
