+++ src/packlets/qrcode/createExplosionRenderer.ts
