+++ src/packlets/qrcode/createPixelsRenderer.ts
